Remove transparency manager
